Add new package manager
This is a rewrite of the package manager, using redo(1) to handle the
dependency resolution, execline as script language, and the venus tools.

It has a quite different approach than the old one, it has atomic
installation/update, easy rollback, support to multiple repositories.

At the current moment it's lacking a frontend binary and providing only
the minimum functionality.

If used along with arbor it should provide an environment with a good
reproducibility.